+++ projects/back-office/src/app/application/pages/subscriptions/components/subscription-modal/subscription-modal.component.ts
@@ -1,16 +1,19 @@
-import {Apollo} from 'apollo-angular';
-import { Component, Inject, OnInit } from '@angular/core';
+import {Apollo, QueryRef} from 'apollo-angular';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-
 import { Application, Channel, Form, Subscription } from '@safe/builder';
-import { Observable } from 'rxjs';
+import { BehaviorSubject, Observable } from 'rxjs';
 import {
   GetRoutingKeysQueryResponse,
   GET_ROUTING_KEYS,
   GET_FORM_NAMES, GetFormsQueryResponse
 } from '../../../../../graphql/queries';
 import { map, startWith } from 'rxjs/operators';
+import { MatSelect } from '@angular/material/select';
+import { MatAutocomplete } from '@angular/material/autocomplete';
+
+const ITEMS_PER_PAGE = 10;
 
 @Component({
   selector: 'app-subscription-modal',
@@ -23,10 +26,29 @@ export class SubscriptionModalComponent implements OnInit {
   subscriptionForm: FormGroup = new FormGroup({});
 
   // === DATA ===
-  public forms: Form[] = [];
+  private forms = new BehaviorSubject<Form[]>([]);
+  public forms$!: Observable<Form[]>;
+  private formsQuery!: QueryRef<GetFormsQueryResponse>;
+  private formsPageInfo = {
+    endCursor: '',
+    hasNextPage: true
+  };
+  private formsLoading = true;
+
+  @ViewChild('formSelect') formSelect?: MatSelect;
+
   // === DATA ===
-  private applications: Application[] = [];
-  public filteredApplications!: Observable<Application[]>;
+  private applications = new BehaviorSubject<Application[]>([]);
+  public filteredApplications$!: Observable<Application[]>;
+  public applications$!: Observable<Application[]>;
+  private applicationsQuery!: QueryRef<GetRoutingKeysQueryResponse>;
+  private applicationsPageInfo = {
+    endCursor: '',
+    hasNextPage: true
+  };
+  private applicationsLoading = true;
+
+  @ViewChild('applicationSelect') applicationSelect?: MatAutocomplete;
 
   get routingKey(): string {
     return this.subscriptionForm.value.routingKey;
@@ -53,26 +75,46 @@ export class SubscriptionModalComponent implements OnInit {
       convertTo: [( this.data.subscription && this.data.subscription.convertTo ) ? this.data.subscription.convertTo.id : ''],
       channel: [( this.data.subscription && this.data.subscription.channel ) ? this.data.subscription.channel.id : '']
     });
-    this.apollo.watchQuery<GetFormsQueryResponse>({
-      query: GET_FORM_NAMES
-    }).valueChanges.subscribe((res: any) => {
-      this.forms = res.data.forms;
+    // Get forms and set pagination logic
+    this.formsQuery = this.apollo.watchQuery<GetFormsQueryResponse>({
+      query: GET_FORM_NAMES,
+      variables: {
+        first: ITEMS_PER_PAGE
+      }
     });
-    this.apollo.watchQuery<GetRoutingKeysQueryResponse>({
-      query: GET_ROUTING_KEYS
-    }).valueChanges.subscribe(res => {
-      this.applications = res.data.applications.filter(x => x.channels ? x.channels.length > 0 : false);
+
+    this.forms$ = this.forms.asObservable();
+    this.formsQuery.valueChanges.subscribe(res => {
+      this.forms.next(res.data.forms.edges.map(x => x.node));
+      this.formsPageInfo = res.data.forms.pageInfo;
+      this.formsLoading = res.loading;
     });
-    this.filteredApplications = this.subscriptionForm.controls.routingKey.valueChanges.pipe(
-      startWith(''),
-      map(value => typeof value === 'string' ? value : value.name),
-      map(x => this.filter(x))
-    );
+
+    // Get applications and set pagination logic
+    this.applicationsQuery = this.apollo.watchQuery<GetRoutingKeysQueryResponse>({
+      query: GET_ROUTING_KEYS,
+      variables: {
+        first: ITEMS_PER_PAGE
+      }
+    });
+
+    // this.applications$ = this.applications.asObservable();
+    this.applicationsQuery.valueChanges.subscribe(res => {
+      this.applications.next(res.data.applications.edges.map(x => x.node).filter(x => x.channels ? x.channels.length > 0 : false));
+      this.applicationsPageInfo = res.data.applications.pageInfo;
+      this.applicationsLoading = res.loading;
+      this.applications$ = this.subscriptionForm.controls.routingKey.valueChanges.pipe(
+        startWith(''),
+        map(value => typeof value === 'string' ? value : value.name),
+        map(x => this.filter(x))
+      );
+    });
   }
 
   private filter(value: string): Application[] {
     const filterValue = value.toLowerCase();
-    return this.applications ? this.applications.filter(x => x.name?.toLowerCase().indexOf(filterValue) === 0) : this.applications;
+    const applications = this.applications.getValue();
+    return applications ? applications.filter(x => x.name?.toLowerCase().indexOf(filterValue) === 0) : applications;
   }
 
   /*  Close the modal without sending any data.
@@ -81,4 +123,83 @@ export class SubscriptionModalComponent implements OnInit {
     this.dialogRef.close();
   }
 
+  /**
+   * Adds scroll listener to select.
+   * @param e open select event.
+   */
+   onOpenFormSelect(e: any): void {
+    if (e && this.formSelect) {
+      const panel = this.formSelect.panel.nativeElement;
+      panel.addEventListener('scroll', (event: any) => this.loadOnScrollForm(event));
+    }
+  }
+
+  /**
+   * Fetches more forms on scroll.
+   * @param e scroll event.
+   */
+  private loadOnScrollForm(e: any): void {
+    if (e.target.scrollHeight - (e.target.clientHeight + e.target.scrollTop) < 50) {
+      if (!this.formsLoading && this.formsPageInfo.hasNextPage) {
+        this.formsLoading = true;
+        this.formsQuery.fetchMore({
+          variables: {
+            first: ITEMS_PER_PAGE,
+            afterCursor: this.formsPageInfo.endCursor
+          },
+          updateQuery: (prev, { fetchMoreResult }) => {
+            if (!fetchMoreResult) { return prev; }
+            return Object.assign({}, prev, {
+              forms: {
+                edges: [...prev.forms.edges, ...fetchMoreResult.forms.edges],
+                pageInfo: fetchMoreResult.forms.pageInfo,
+                totalCount: fetchMoreResult.forms.totalCount
+              }
+            });
+          }
+        });
+      }
+    }
+  }
+
+  /**
+   * Adds scroll listener to auto complete.
+   */
+   onOpenApplicationSelect(): void {
+    if (this.applicationSelect) {
+      const panel = this.applicationSelect.panel.nativeElement;
+      if (panel) {
+        panel.onscroll = (event: any) => this.loadOnScrollApplication(event);
+      }
+    }
+  }
+
+  /**
+   * Fetches more forms on scroll.
+   * @param e scroll event.
+   */
+  private loadOnScrollApplication(e: any): void {
+    if (e.target.scrollHeight - (e.target.clientHeight + e.target.scrollTop) < 50) {
+      console.log(this.applicationsPageInfo.hasNextPage);
+      if (!this.applicationsLoading && this.applicationsPageInfo.hasNextPage) {
+        this.applicationsLoading = true;
+        this.applicationsQuery.fetchMore({
+          variables: {
+            first: ITEMS_PER_PAGE,
+            afterCursor: this.applicationsPageInfo.endCursor
+          },
+          updateQuery: (prev, { fetchMoreResult }) => {
+            if (!fetchMoreResult) { return prev; }
+            return Object.assign({}, prev, {
+              applications: {
+                edges: [...prev.applications.edges, ...fetchMoreResult.applications.edges],
+                pageInfo: fetchMoreResult.applications.pageInfo,
+                totalCount: fetchMoreResult.applications.totalCount
+              }
+            });
+          }
+        });
+      }
+    }
+  }
 }