+++ projects/back-office/src/app/dashboard/pages/update-record/update-record.component.ts
@@ -2,8 +2,8 @@ import {Apollo} from 'apollo-angular';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { GetRecordByIdQueryResponse, GET_RECORD_BY_ID } from '../../../graphql/queries';
-import { Record } from '@safe/builder';
+import { GetFormByIdQueryResponse, GetRecordByIdQueryResponse, GET_FORM_STRUCTURE, GET_RECORD_BY_ID } from '../../../graphql/queries';
+import { Record, Form } from '@safe/builder';
 
 @Component({
   selector: 'app-update-record',
@@ -15,7 +15,8 @@ export class UpdateRecordComponent implements OnInit {
   // === DATA ===
   public loading = true;
   public id = '';
-  public record: Record = {};
+  public record?: Record;
+  public form?: Form;
   public backPath = '';
 
   constructor(
@@ -27,6 +28,18 @@ export class UpdateRecordComponent implements OnInit {
   ngOnInit(): void {
     this.id = this.route.snapshot.paramMap.get('id') || '';
     this.backPath = this.router.url.replace(`/update/${this.id}`, '');
+    const template = history.state.template;
+    if (template) {
+      this.apollo.watchQuery<GetFormByIdQueryResponse>({
+        query: GET_FORM_STRUCTURE,
+        variables: {
+          id: template
+        }
+      }).valueChanges.subscribe(res => {
+        this.form = res.data.form;
+        this.loading = res.loading;
+      });
+    }
     if (this.id !== null) {
       this.apollo.watchQuery<GetRecordByIdQueryResponse>({
         query: GET_RECORD_BY_ID,
@@ -35,7 +48,10 @@ export class UpdateRecordComponent implements OnInit {
         }
       }).valueChanges.subscribe(res => {
         this.record = res.data.record;
-        this.loading = res.loading;
+        if (!template) {
+          this.form = this.record.form || {};
+          this.loading = res.loading;
+        }
       });
     }
   }